Backend/TravelQuest/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
django_filter==24.2
python-dotenv==1.0.1
Pillow==10.3.0
psycopg2-binary==2.9.9
six==1.16.0
